% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_function}
\alias{get_function}
\title{Access the original function from a rate limited function}
\usage{
get_function(f)
}
\arguments{
\item{f}{A rate limited function or group of functions}
}
\description{
Access the original function from a rate limited function
}
