% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_precision}
\alias{get_precision}
\title{Access the rate limit precision}
\usage{
get_precision(f)
}
\arguments{
\item{f}{A rate limited function or group of functions}
}
\description{
Access the rate limit precision
}
