% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_rates}
\alias{get_rates}
\title{Access the rate limit(s) of a rate limited function}
\usage{
get_rates(f)
}
\arguments{
\item{f}{A rate limited function or group of functions}
}
\description{
Access the rate limit(s) of a rate limited function
}
