% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-dispenser.R
\name{request}
\alias{request}
\alias{request.token_dispenser}
\title{Request a token from a token dispenser}
\usage{
request(x)

\method{request}{token_dispenser}(x)
}
\arguments{
\item{x}{A \code{\link{token_dispenser}}}
}
\value{
TRUE (possibly after a delay)
}
\description{
Once you've created a \code{\link{token_dispenser}}, use this function to
request tokens. Tokens will be disbursed subject to the rate limit implied
by the \code{token_dispenser}.
}

