% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit-rate.R
\name{reset}
\alias{reset}
\title{Re-create a rate-limited function}
\usage{
reset(f)
}
\arguments{
\item{f}{A rate-limited functoin}
}
\description{
This function does not modify the original rate-limited function, instead
it returns a new function with the same rate limits (but no memory of prior
function calls).
}

