% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-dispenser.R
\name{token_dispenser}
\alias{token_dispenser}
\title{Create a new token dispenser}
\usage{
token_dispenser(n, period, precision = 60)
}
\arguments{
\item{n}{Number of tokens disbursed in the given time period}

\item{period}{Length of period in seconds}

\item{precision}{Resolution with which to measure time, in fractions of a second}
}
\description{
A token dispenser is an object with one method, \code{\link{request}}, which
always returns TRUE, possibly after a delay. This function initializes a
token dispenser with the given parameters. A token dispenser will never
disburse more than \code{n} tokens in any window of length \code{period}.
}
\details{
In order to guarantee compliance with a rate limit, the token dispenser
by default measures time intervals to the 1/60th of a second, always rounding
down. Use the \code{precision} argument to specify what fraction of a second
to use (60 means 1/60th, 100 means 1/100, etc).
}

