% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit-rate.R
\name{limit_rate}
\alias{limit_rate}
\title{Limit the rate at which a function will execute}
\usage{
limit_rate(f, ..., precision = 60)
}
\arguments{
\item{f}{The function to be rate-limited}

\item{...}{One or more rates, created using \code{\link{rate}}}
}
\description{
Limit the rate at which a function will execute
}

